+++ apps/web/src/chat/ChatSurface.tsx
@@ -1915,7 +1915,7 @@ function ChatSessionSlot({
                   {a.icon}
                 </Button>
               ))}
-              <ComposerModelSelect />
+              <ComposerModelSelect onRuntimeSwitch={(note) => noteToThread(note, { tone: "info" })} />
               {session?.incognito ? (
                 <button
                   type="button"